Refactor how test database is made.
In the old system, it would create a totally new database each time a
test was run. This is nice because it ensure that the database is
always empty each time a test is run. This make sure no conflicts
happen (or duplicate data is created). And because we're running is a
dockr container we, as the maintainers of the databaser, won't really
notice all these superfluous database being created because when docker
or our machine restarts they all get washed away.

This has the downside in that we can't query our test database because
it has a random name. So now we have chosen a static name, test_skills,
and always create and recreate that database.

There is still some work to be done in that we need to refactor our
configuration file to contain the idea of dev and test
configurations. The configuration file might look something like:

```
dev:
  application_port: 8000
  database:
    host: "localhost"
    port: 5432
    username: "postgres"
    password: "password"
    database_name: "skills"
test:
  application_port: 8000
  database:
    host: "localhost"
    port: 5432
    username: "postgres"
    password: "password"
    database_name: "test_skills"
```